Create new namespace for RecIM allowing for 'duplicate' service renaming
