+++ Gordon360/Services/ServiceInterfaces.cs
@@ -12,6 +12,19 @@ using static Gordon360.Services.MembershipService;
 // Namespace with all the Service Interfaces that are to be implemented. I don't think making this interface is required, the services can work fine on their own.
 // However, building the interfaces first does give a general sense of structure to their implementations. A certain cohesiveness :p.
 // </summary>
+namespace Gordon360.Services.RecIM
+{
+    public interface IActivityService
+    {
+        IEnumerable<Activity> GetActivities();
+        Activity GetActivityByID(int ActivityID);
+        IEnumerable<Activity> GetActivitiesByTime(DateTime? time);
+        Task UpdateActivity(Activity updatedActivity);
+        Task PostActivity(Activity newActivity);
+
+    }
+}
+
 namespace Gordon360.Services
 {
     public interface IProfileService
@@ -34,16 +47,6 @@ namespace Gordon360.Services
         Task UpdateProfileImageAsync(string username, string path, string name);
         ProfileViewModel? ComposeProfile(object? student, object? alumni, object? faculty, object? customInfo);
         Task InformationChangeRequest(string username, ProfileFieldViewModel[] updatedField);
-    }
-
-    public interface IRecIMActivityService
-    {
-        IEnumerable<Activity> GetActivities();
-        Activity GetActivityByID(int ActivityID);
-        IEnumerable<Activity> GetActivitiesByTime(DateTime? time);
-        Task UpdateActivity(Activity updatedActivity);
-        Task PostActivity(Activity newActivity);
-    
     }
 
     public interface ISeriesService
